Final game code fixes (see description)
1) Fixed all warning messages (conversion or truncate);
2) Added some translation phrases;
3) Improved AudioManager class;

+++ centurion-source/source/audio/recorder.cpp
@@ -36,7 +36,7 @@ ISoundBuffer *Recorder::CopyToBuffer()
 		SoundBuffer* buff = new SoundBuffer();
 		buff->SetSampleRate(44100);
 		buff->WriteData(&samples[0], samples.size());
-		buff->SetSampleCount(samples.size());
+		buff->SetSampleCount((unsigned int)samples.size());
 		samples.resize(0);
 		samples.clear();
 		recorderMutex.unlock();
