+++ centurion-source/source/menu/menu.cpp
@@ -42,8 +42,8 @@ void Menu::create(std::vector<Player> *List) {
 
 			/* use data */
 			img = gui::Image(imageName);	
-			if (size == "auto") img.create(align, x, y, 0, 0, pickingId * clickable);
-			if (size == "max") img.create(align, x, y, getParam("window-width"), getParam("window-height"), pickingId * clickable);
+			if (size == "auto") img.create(align, (float)x, (float)y, 0.0f, 0.0f, pickingId * clickable);
+			if (size == "max") img.create(align, (float)x, (float)y, getParam("window-width"), getParam("window-height"), pickingId * clickable);
 			images[menus[a]].push_back(img);
 
 			// update picking
