+++ centurion-source/source/unit/unit_functions.cpp
@@ -38,20 +38,20 @@ bool unit::isInSelectionRect(std::array<float, 8> &coords){
 }
 
 float unit::getAngle(std::vector<glm::ivec2> &path, int &n) {
-	float angle = atan2(path[n + 1].y - path[n].y, path[n + 1].x - path[n].x) * 180 / 3.14159265;
+	float angle = (float)atan2(path[n + 1].y - path[n].y, path[n + 1].x - path[n].x) * 180.0f / 3.14159265f;
 	if (angle < 0) angle += 360.0f;
 	return angle;
 }
 
 float unit::getDistance(std::vector<glm::ivec2> &path, int &n) {
-	float deltaX = path[n+1].x - path[n].x;
-	float deltaY = path[n+1].y - path[n].y;
+	float deltaX = (float)path[n+1].x - (float)path[n].x;
+	float deltaY = (float)path[n+1].y - (float)path[n].y;
 	return sqrt(deltaX * deltaX + deltaY * deltaY);
 }
 
 float unit::getResDistance(std::vector<glm::ivec2> &path, int &n, glm::vec3 &pos2d) {
-	float deltaX = path[n+1].x - (int)pos2d.x / PATH::CELL_GRID_SIZE * PATH::CELL_GRID_SIZE;
-	float deltaY = path[n+1].y - (int)pos2d.y / PATH::CELL_GRID_SIZE * PATH::CELL_GRID_SIZE;
+	float deltaX = (float)path[n+1].x - (int)pos2d.x / PATH::CELL_GRID_SIZE * PATH::CELL_GRID_SIZE;
+	float deltaY = (float)path[n+1].y - (int)pos2d.y / PATH::CELL_GRID_SIZE * PATH::CELL_GRID_SIZE;
 	return sqrt(deltaX * deltaX + deltaY * deltaY);
 }
